src/apps/nfc/tag/controllers/find-tag.controller.ts
import { Controller, Get, Body } from '@nestjs/common';
import { ApiTags, ApiOkResponse, ApiOperation, ApiBody, ApiQuery } from '@nestjs/swagger';
import { TagDto } from './../dto/tag.dto';

// @hades
import { IQueryBus } from '@hades/shared/domain/bus/query-bus';
import { QueryStatementInput } from '@hades/shared/domain/persistence/sql-statement-input';
import { FindTagQuery } from '@hades/nfc/tag/application/find/find-tag.query';

@ApiTags('[nfc] tag')
@Controller('nfc/tag')
export class FindTagController 
{
    constructor(
        private readonly queryBus: IQueryBus
    ) {}

    @Get()
    @ApiOperation({ summary: 'Find tag according to query' })
    @ApiOkResponse({ description: 'The record has been successfully created.', type: TagDto })
    @ApiBody({ type: [QueryStatementInput] })
    @ApiQuery({ name: 'query', type: [QueryStatementInput] })
    async main(@Body('query') queryStatements: QueryStatementInput[])
    {
        return await this.queryBus.ask(new FindTagQuery(queryStatements));   
    }
}